Minor refactoring for authenticaton
Summary:
Caching of creds is an orthogonal concern, encapsulate it in a
decorator and do not let caching semantics spread themselves out
