+++ app/collins/util/security/FileAuthenticationProvider.scala
@@ -13,7 +13,7 @@ class FileAuthenticationProvider() extends AuthenticationProvider {
   def userfile = FileAuthenticationProviderConfig.userfile
   override val authType = Array("file")
 
-  lazy private val userCache = ConfigCache.create(10000L, FileUserLoader())
+  private val userCache = ConfigCache.create(10000L, FileUserLoader())
 
   override def authenticate(username: String, password: String): Option[User] = {
     user(username) match {
